+++ test/unit-tests/payableProxyController.test.ts
@@ -10,9 +10,9 @@ import {
   AddressBookInstance,
   OwnedUpgradeabilityProxyInstance,
   PayableProxyControllerInstance,
-} from '../build/types/truffle-types'
+} from '../../build/types/truffle-types'
 import BigNumber from 'bignumber.js'
-import {createTokenAmount, createScaledNumber} from './utils'
+import {createTokenAmount, createScaledNumber} from '../utils'
 
 const {expectRevert, time} = require('@openzeppelin/test-helpers')
